chore: move mappers and DTOs to application layer
