+++ src/core/application/use-cases/mappers/mapper.ts
@@ -1,0 +1,10 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+export abstract class Mapper<DTO, Domain, Persistence> {
+  protected constructor() {}
+  // abstract toDTO(domain: Domain): DTO
+  // abstract toCollectionDTO(domainCollection: Domain[]): DTO[]
+  // abstract toDomain(persistence: Persistence): Domain
+  // abstract toCollectionDomain(persistenceCollection: Persistence[]): Domain[]
+  // abstract toPersistence(domain: Domain): Persistence
+  // abstract toCollectionPersistence(domainCollection: Domain[]): Persistence[]
+}
