--- src/core/contracts/mapper.ts
@@ -1,10 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-export abstract class Mapper<DTO, Entity, Persistence> {
-  protected constructor() {}
-  // abstract toDTO(entity: Entity): DTO
-  // abstract toCollectionDTO(entityCollection: Entity[]): DTO[]
-  // abstract toEntity(persistence: Persistence): Entity
-  // abstract toCollectionEntity(persistenceCollection: Persistence[]): Entity[]
-  // abstract toPersistence(entity: Entity): Persistence
-  // abstract toCollectionPersistence(entityCollection: Entity[]): Persistence[]
-}
