+++ src/modules/users/application/use-cases/fetch-users-use-case.ts
@@ -3,8 +3,8 @@ import { type UseCase } from '@/core/application/use-cases/use-case'
 import { type PaginationParams } from '@/core/shared/types/pagination-params'
 import { type CountUsersRepository } from '@/modules/users/application/repositories/count-users-repository'
 import { type FindManyUsersRepository } from '@/modules/users/application/repositories/find-many-users-repository'
-import { type UserCollectionDTO } from '@/modules/users/contracts/dtos/user-dto'
-import { UserMap } from '@/modules/users/contracts/mappers/user-map'
+import { type UserCollectionDTO } from '@/modules/users/application/use-cases/dtos/user-dto'
+import { UserMapper } from '@/modules/users/application/use-cases/mappers/user-mapper'
 
 export namespace FetchUsers {
   export interface Input {
@@ -37,7 +37,7 @@ export class FetchUsersUseCase
     ])
     return right({
       count,
-      users: UserMap.toCollectionDTO(users),
+      users: UserMapper.toCollectionDTO(users),
     })
   }
 }
