+++ src/modules/users/application/use-cases/get-user-by-id-use-case.ts
@@ -2,8 +2,8 @@ import { type Either, left, right } from '@/core/application/either'
 import { type UseCase } from '@/core/application/use-cases/use-case'
 import { UserNotFoundError } from '@/modules/users/application/errors/user-not-found-error'
 import { type FindUserByIdRepository } from '@/modules/users/application/repositories/find-user-by-id-repository'
-import { type UserDTO } from '@/modules/users/contracts/dtos/user-dto'
-import { UserMap } from '@/modules/users/contracts/mappers/user-map'
+import { type UserDTO } from '@/modules/users/application/use-cases/dtos/user-dto'
+import { UserMapper } from '@/modules/users/application/use-cases/mappers/user-mapper'
 
 export namespace GetUserById {
   export interface Input {
@@ -31,7 +31,7 @@ export class GetUserByIdUseCase
       return left(new UserNotFoundError(id))
     }
     return right({
-      user: UserMap.toDTO(user),
+      user: UserMapper.toDTO(user),
     })
   }
 }
