+++ src/modules/users/infra/repositories/find-user-by-id-prisma-repository.ts
@@ -1,6 +1,6 @@
 import { BasePrismaRepository } from '@/core/infra/repositories/base-prisma-repository'
 import { type FindUserByIdRepository } from '@/modules/users/application/repositories/find-user-by-id-repository'
-import { UserMap } from '@/modules/users/contracts/mappers/user-map'
+import { UserMapper } from '@/modules/users/application/use-cases/mappers/user-mapper'
 import { type UserEntity } from '@/modules/users/domain/entities/user-entity'
 
 export class FindUserByIdPrismaRepository
@@ -20,6 +20,6 @@ export class FindUserByIdPrismaRepository
     if (!user) {
       return null
     }
-    return UserMap.toEntity(user)
+    return UserMapper.toDomain(user)
   }
 }
